Remove invalid package from example_packages
This was causing problems in Travis because it was deliberately awful Go code
