--- example_packages/invalid/main.go
@@ -1,9 +1,0 @@
-package invalid
-
-import (
-	"log"
-
-
-funk main() {
-	return e
-)
